Finalizado el ajustes del BUG
