--- src/test/java/co/vinni/itsdna/ItsadnApplicationTests.java
@@ -1,22 +1,0 @@
-package co.vinni.itsdna;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-@SpringBootTest
-class ItsdnaApplicationTests {
-
-    @Test
-    void contextLoads() {
-        assertEquals(2,2);
-    }
-
-    @Test
-    void contextLoads1() {
-        assertEquals(2,2);
-    }
-
-
-}
